Split Nodes to separate files and add more TypeScript
Add TypeScript types and split nodes into separate files.

* Import the generated types from `src/workflow.d.ts` in `src/components/GraphEditor.vue`.
* Use the generated types for `loadWorkflow`, `nodeData`, `linkData`, and `config` in `addNode` function in `src/components/GraphEditor.vue`.
* Move node class definitions from `src/nodes/nodes.js` to separate files: `llmNode.ts`, `textToSpeechNode.ts`, `audioInputNode.ts`, `audioInputTransportNode.ts`, `audioOutputTransportNode.ts`, `sttNode.ts`, `ttsNode.ts`, and `textNode.ts`.
* Create `src/nodes/nodes.ts` to import and register all node definitions.
* Delete `src/nodes/nodes.js`.
* Create `src/workflow.d.ts` manually based on `workflow.json` file with generic `Workflow`, `NodeData`, and `LinkData` types.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/LucasMatuszewski/housecat-LowCode-Pipecat?shareId=XXXX-XXXX-XXXX-XXXX).
