+++ src/components/GraphEditor.vue
@@ -16,6 +16,7 @@ import { LGraph, LGraphCanvas, LiteGraph, LGraphNode } from '@comfyorg/litegraph
 import { useMainStore } from '../pinia';
 import '../nodes/nodes.js';  // Import the custom nodes
 import workflowData from '../workflow.json';  // Import the workflow JSON
+import { Workflow, NodeData, LinkData } from '../workflow';  // Import the workflow types
 
 export default defineComponent({
   name: 'GraphEditor',
@@ -76,13 +77,13 @@ export default defineComponent({
       });
     });
 
-    const loadWorkflow = (data: any) => {
+    const loadWorkflow = (data: Workflow) => {
       console.log("_____GraphEditor.vue loadWorkflow")
       graph.clear();
       
       // Create nodes
-      const nodeMap = new Map();
-      data.nodes.forEach((nodeData: any) => {
+      const nodeMap = new Map<number, LGraphNode>();
+      data.nodes.forEach((nodeData: NodeData) => {
         const node = LiteGraph.createNode(nodeData.type);
         if (node) {
           node.configure(nodeData);
@@ -94,7 +95,7 @@ export default defineComponent({
       });
 
       // Create links
-      data.links.forEach((linkData: any) => {
+      data.links.forEach((linkData: LinkData) => {
         const [, originNodeId, originSlot, targetNodeId, targetSlot] = linkData;
         const originNode = nodeMap.get(originNodeId);
         const targetNode = nodeMap.get(targetNodeId);
@@ -109,21 +110,20 @@ export default defineComponent({
       graph.setDirtyCanvas(true, true);
     };
 
-    const exportWorkflow = (graph) => {
+    const exportWorkflow = (graph: LGraph): Workflow => {
       console.log("_____GraphEditor.vue exportWorkflow")
-      let linksArr = []
-      let theseNodes = []
+      let linksArr: LinkData[] = []
+      let theseNodes: NodeData[] = []
       for (let nn in graph.nodes) {
         let n = graph.nodes[nn];
-        let tmp = {};
-        let asdf = JSON.stringify(n.properties);
-        tmp.id = n.id;
-        tmp.inputs = n.inputs;
-        tmp.pos = n.pos;
-        tmp.properties = n.properties;
-        tmp.type = n.type;
+        let tmp: NodeData = {
+          id: n.id,
+          inputs: n.inputs,
+          pos: n.pos,
+          properties: n.properties,
+          type: n.type
+        };
 
-        // const [, originNodeId, originSlot, targetNodeId, targetSlot] = linkData;
         if (graph.links[nn]) {
           linksArr.push([n.id,
               graph.links[nn].origin_id,
@@ -137,7 +137,7 @@ export default defineComponent({
         theseNodes.push(tmp);
       }
 
-      let outputJson = {
+      let outputJson: Workflow = {
         "nodes": [],
         "links": [],
         "groups": [],
@@ -147,12 +147,11 @@ export default defineComponent({
       outputJson.nodes = theseNodes;
       outputJson.links = linksArr;
 
-      // console.log("_____GraphEditor.vue exportWorkflow outputJson:", outputJson)
       return outputJson;
     };
 
 
-    const addNode = (type: string, config: any = {}) => {
+    const addNode = (type: string, config: Partial<NodeData> = {}) => {
       const node = LiteGraph.createNode(type);
       if (node) {
         node.configure(config);
