+++ src/nodes/textToSpeechNode.ts
@@ -1,0 +1,19 @@
+import { LiteGraph } from '@comfyorg/litegraph';
+
+class TextToSpeechNode extends LiteGraph.LGraphNode {
+    constructor() {
+        super();
+        this.title = "Text to Speech";
+        this.addInput("text", "string");
+        this.addOutput("audio", "audio");
+        this.color = "rgb(100,100,10)"
+    }
+
+    onExecute() {
+        const text = this.getInputData(0);
+        // Implement TTS logic here
+        this.setOutputData(0, { type: "audio", data: "synthesized_audio_data" });
+    }
+}
+
+export { TextToSpeechNode };
